+++ src/prompt.ts
@@ -1,11 +1,13 @@
+import { PathDetails } from '@r35007/mock-server/dist/server/types/common.types';
 import * as vscode from "vscode";
 import { PromptAction } from './enum';
 import { Settings } from "./Settings";
 
 export class Prompt {
-  static getEnvironment = async (envNameList: string[]) => {
-    return vscode.window.showQuickPick(envNameList, {
-      placeHolder: "Please select any environment",
+  static getEnvironment = async (envNameList: PathDetails[]) => {
+    const pickList = envNameList.map(env => ({ ...env, label: env.fileName, description: env.filePath }));
+    return vscode.window.showQuickPick(pickList, {
+      placeHolder: "Please select a file",
     });
   };